Unbox `automaton` submodule and get rid of `acceptor.rs` within
